FAIR-68 minor alert refactors & added alert to rootReducer
