+++ src/components/alert/AlertComponents.css
@@ -1,0 +1,95 @@
+.fairwear-banner-title {
+	width: 100%;
+	margin: 0 !important;
+}
+
+.fairwear-banner {
+	display: flex;
+	align-items: center;
+	justify-content: flex-start;
+	border-radius: 0;
+	padding: 8px 32px !important;
+}
+
+.fairwear-banner .content {
+	display: flex;
+	width: 100%;
+	gap: 8px;
+}
+
+.fairwear-banner .content .text-and-logo-container {
+	display: flex;
+	align-items: center;
+	justify-content: flex-start;
+	width: 100%;
+}
+
+.fairwear-banner .content .action-button-container {
+	display: flex;
+	align-items: flex-start;
+	justify-content: flex-end;
+	gap: 16px;
+	width: 100%;
+}
+
+.fairwear-toast .content .action-button-container {
+	display: flex;
+	align-items: center;
+	justify-content: flex-end;
+	gap: 16px;
+	width: 25%;
+}
+
+.fairwear-banner-action-button {
+	padding: 8px 16px !important;
+	text-transform: none !important;
+	box-shadow: none !important;
+}
+
+.fairwear-banner-close-button {
+	padding: 8px !important;
+	text-transform: none !important;
+}
+
+.fairwear-toast {
+	display: flex;
+	align-items: center;
+	justify-content: flex-start;
+	border-radius: 0 !important;
+	padding: 6px 16px !important;
+	height: 48px;
+}
+
+.fairwear-toast .toast-alert {
+	display: flex;
+	align-items: center;
+	justify-content: flex-start;
+	width: 100%;
+	min-width: 260px;
+	padding: 0 16px !important;
+	gap: 8px;
+	border-radius: 4px;
+}
+
+.toast-alert .content {
+	display: flex;
+	align-items: center;
+	width: 100%;
+	gap: 8px;
+}
+
+.fairwear-toast .content .text-and-logo-container {
+	display: flex;
+	align-items: center;
+	justify-content: flex-start;
+	width: 75%;
+}
+
+@media (max-width: 450px) {
+	.fairwear-banner .content {
+		flex-direction: column;
+	}
+	.fairwear-banner .content .action-button-container {
+		justify-content: left;
+	}
+}
